Render a welcome message after users have set up their account
WorkOS redirects back to the default redirect URL after a user
finishes setup.  In this case, we don't need to handle the code
exchange like we would in a typical OAuth dance because the user
is not actually being authenticated for the gateway application.
Instead, the MCP client will handle the OAuth dance when the user
uses Ragie via MCP.